requirements/requirements_conda.txt
jupyterlab
numpy
scikit-learn
yellowbrick
folium
geopy
fastapi
mkdocs
mkdocstrings-python
mkdocs-material
sqlalchemy
psycopg2-binary
jupysql
pandas
mlflow
boto3
pyarrow
plotly
ipykernel
python-dotenv
pytorch
faiss-cpu